Rebuild this change on top of main (#669-#672) instead of merging the old
diff textually. Kept, each checked against retail or the comparison:

- WorldSetCameraLocation takes const srVector3T<float>*: its callers pass
  vectors, so the float* reinterpret and the srNode C-casts go away.
- No parameter or local slot doubles as a FileRead/FileWrite count
  (SaveFactState, LoadFactState, LoadMonsterDatabaseRecord,
  LoadMonsterDatabaseRange, W8VirtualFileBinIStream::vread, SaveItemFile);
  each takes an ordinary local, and the copies that existed only to allow
  the aliasing are dropped. SaveItemFile and LoadMonsterDatabaseRange
  compare closer to retail.
- UpdateUseItemSelect calls CommitSelectedItemUse instead of spelling it out.
- EnableAllRenderOptions is a counted for loop.
- The C-style cast gate ignores cast spellings inside string literals, and
  the decompiler-quality report records the canonical program name.

Dropped: the octree/bounds, colour-vector, GrCycle, NPC-interaction,
stModelInstance and Monster edits (in flight elsewhere); goto rewrites; the
uninitialized-read policy change, its CMake -Wno-error list and the
MATCH-uninitialized locals (a policy decision, not a cleanup); the
PList/IList and text-box-macro de-inlining (regresses six exact functions);
quad.cpp's minimum_y change (retail subtracts minimum x); the reviewed
Ghidra checkpoint refresh; and deletions of evidence comments.

+++ src/wiz8/engine_code/IntervalGate.cpp
@@ -82,8 +82,8 @@ BOOLEAN W8IntervalGate::Save(int handle)
     float elapsed;
     if ((m_flags & 2) == 0) {
         int sample = ReadClock();
-        float progress =
-            (float)(unsigned int)(sample - m_start) / (float)(unsigned int)(m_end - m_start);
+        float progress = static_cast<unsigned int>(sample - m_start) /
+                         static_cast<float>(static_cast<unsigned int>(m_end - m_start));
         // The on-disk pair and its OR-combined result mirror W8GameTimer::Load.
         BOOLEAN saved = FileWrite(handle, &progress, sizeof(progress), 0);
         saved |= FileWrite(handle, &m_duration_scale, sizeof(m_duration_scale), 0);
